Add ability to edit steps
I completed the put request to edit the text of a specific step. I also modified the post request to hand inserting a step in the middle of a set of existing steps. If a step is posted with an existing num, all steps with a num greater than or equal to the new number get shifted up by one.